Screenshot mode.
A new user input has been enabled to take screenshots in high
resolution. Both the simulator and the viewer can trigger this
behavior. The auto-input can also specify it and change the
camera parameters to take it from the correct perspective.

The viewer also has a new screenshot mode which skips to certain
frames specified by tags and takes appropriate screenshots.

Note that, just as with exporting the sidebar video, it requires a
graphics context, so it is not available in headless mode.

+++ mono-rfs/Program.cs
@@ -93,37 +93,39 @@ public class Program
 	//[STAThread]
 	public static void Main(string[] args)
 	{
-		string recfile       = "data.zip";
-		string scenefile     = "";
-		string commandfile   = "";
-		string configfile    = "";
-		int    particlecount = 1;
-		bool   onlymapping   = false;
-		bool   realtime      = false;
-		bool   viewer        = false;
-		bool   filterhistory = false;
-		bool   headless      = false;
-		bool   noterminate   = false;
-		bool   showhelp      = false;
+		string recfile        = "data.zip";
+		string scenefile      = "";
+		string commandfile    = "";
+		string configfile     = "";
+		int    particlecount  = 1;
+		bool   onlymapping    = false;
+		bool   realtime       = false;
+		bool   viewer         = false;
+		bool   filterhistory  = false;
+		bool   screenshotmode = false;
+		bool   headless       = false;
+		bool   noterminate    = false;
+		bool   showhelp       = false;
 
 		VehicleType         input     = VehicleType.Simulation;
 		NavigationAlgorithm algorithm = NavigationAlgorithm.PHD;
 
 		OptionSet options = new OptionSet {
-			{ "f|scene=",      "Scene description file. Simulated, recorded or device id.",     f       => scenefile     = f },
-			{ "r|recfile=",    "Recording file. Saves State and events for reviewing.",         r       => recfile       = r },
-			{ "c|command=",    "Auto-command file (simulates user input).",                     c       => commandfile   = c },
-			{ "g|config=",     "Configuration file. Contains global constants",                 g       => configfile    = g },
-			{ "a|algorithm=",  "SLAM solver algorithm ('odometry', 'phd', 'loopy' or 'isam2')", a       => algorithm     = (a == "isam2") ? NavigationAlgorithm.ISAM2 : (a == "odometry") ? NavigationAlgorithm.Odometry : (a == "loopy") ? NavigationAlgorithm.LoopyPHD : NavigationAlgorithm.PHD },
-			{ "p|particles=",  "Number of particles used for the RB-PHD",                       (int p) => particlecount = p },
-			{ "y|onlymap",     "Only do mapping, assuming known localization.",                 y       => onlymapping   = y != null },
-			{ "i|input=",      "Vehicle input stream: 'kinect', 'simulation' or 'record'",      i       => input         = (i == "kinect") ? VehicleType.Kinect : (i == "record") ? VehicleType.Record : VehicleType.Simulation },
-			{ "R|realtime",    "Process the system in realtime, instead of a fixed step.",      R       => realtime      = R != null },
-			{ "v|view",        "View a precorded session.",                                     v       => viewer        = v != null },
-			{ "H|history=",    "Trajectory history mode: either 'filter' or 'smooth'",          h       => filterhistory = (h == "filter") },
-			{ "x|headless",    "Run headless, i.e. with no GUI",                                x       => headless      = x != null },
-			{ "t|noterminate", "Skip simulation termination due command depletion o similars",  t       => noterminate   = t != null },
-			{ "h|help",        "Show this message and exit",                                    h       => showhelp      = h != null }
+			{ "f|scene=",      "Scene description file. Simulated, recorded or device id.",      f       => scenefile      = f },
+			{ "r|recfile=",    "Recording file. Saves State and events for reviewing.",          r       => recfile        = r },
+			{ "c|command=",    "Auto-command file (simulates user input).",                      c       => commandfile    = c },
+			{ "g|config=",     "Configuration file. Contains global constants.",                 g       => configfile     = g },
+			{ "a|algorithm=",  "SLAM solver algorithm ('odometry', 'phd', 'loopy' or 'isam2').", a       => algorithm      = (a == "isam2") ? NavigationAlgorithm.ISAM2 : (a == "odometry") ? NavigationAlgorithm.Odometry : (a == "loopy") ? NavigationAlgorithm.LoopyPHD : NavigationAlgorithm.PHD },
+			{ "p|particles=",  "Number of particles used for the RB-PHD.",                       (int p) => particlecount  = p },
+			{ "y|onlymap",     "Only do mapping, assuming known localization.",                  y       => onlymapping    = y != null },
+			{ "i|input=",      "Vehicle input stream: 'kinect', 'simulation' or 'record'.",      i       => input          = (i == "kinect") ? VehicleType.Kinect : (i == "record") ? VehicleType.Record : VehicleType.Simulation },
+			{ "R|realtime",    "Process the system in realtime, instead of a fixed step.",       R       => realtime       = R != null },
+			{ "v|view",        "View a precorded session.",                                      v       => viewer         = v != null },
+			{ "H|history=",    "Trajectory history mode: either 'filter' or 'smooth'.",          h       => filterhistory  = (h == "filter") },
+			{ "s|screenshot",  "Screenshot mode: just take the screenshots in the tags.",        s       => screenshotmode = s != null },
+			{ "x|headless",    "Run headless, i.e. with no GUI.",                                x       => headless       = x != null },
+			{ "t|noterminate", "Skip simulation termination due command depletion o similars.",  t       => noterminate    = t != null },
+			{ "h|help",        "Show this message and exit.",                                    h       => showhelp       = h != null }
 		};
 
 		try {
@@ -165,7 +167,7 @@ public class Program
 			Run<Linear2DMeasurer, LinearPose2D, LinearMeasurement2D>(
 			    recfile, scenefile, commandfile,
 			    particlecount, onlymapping, realtime, viewer,
-			    filterhistory, headless, noterminate,
+			    filterhistory, screenshotmode, headless, noterminate,
 			    input, algorithm);
 			break;
 		case DynamicsModel.PRM3D:
@@ -174,14 +176,14 @@ public class Program
 			Run<KinectMeasurer, Pose3D, PixelRangeMeasurement>(
 			    recfile, scenefile, commandfile,
 			    particlecount, onlymapping, realtime, viewer,
-			    filterhistory, headless, noterminate,
+			    filterhistory, screenshotmode, headless, noterminate,
 				input, algorithm);
 			}
 			else {
 			Run<PRM3DMeasurer, Pose3D, PixelRangeMeasurement>(
 			    recfile, scenefile, commandfile,
 			    particlecount, onlymapping, realtime, viewer,
-			    filterhistory, headless, noterminate,
+			    filterhistory, screenshotmode, headless, noterminate,
 				input, algorithm);
 			}
 			break;
@@ -202,7 +204,7 @@ public class Program
 	public static void Run<MeasurerT, PoseT, MeasurementT>(
 	                       string recfile, string scenefile, string commandfile,
 	                       int particlecount, bool onlymapping, bool realtime, bool viewer,
-	                       bool filterhistory, bool headless, bool noterminate,
+	                       bool filterhistory, bool screenshotmode, bool headless, bool noterminate,
 	                       VehicleType input, NavigationAlgorithm algorithm)
 		where PoseT        : IPose<PoseT>, new()
 		where MeasurementT : IMeasurement<MeasurementT>, new()
@@ -213,37 +215,47 @@ public class Program
 			string tmpdir;
 
 			using (var sim = Viewer<MeasurerT, PoseT, MeasurementT>.
-			                     FromFiles(recfile, filterhistory, out tmpdir)) {
+			                     FromFiles(recfile, filterhistory, screenshotmode, out tmpdir)) {
+				sim.ScreenshotPrefix = recfile + "-screenshot-";
 				setSignalHandler(sim);
-				sim.Run();
 
-				if (sim.TagChanged) {
-					string datadir = Path.Combine(tmpdir, "data");
-					string bkpfile = recfile + ".bk";
+				if (sim.Estimate.Count > 0) {
+					sim.Run();
 
-					Console.WriteLine("  -- tags were modified, rewriting");
-					File.WriteAllText(Path.Combine(datadir, "tags.out"), sim.SerializedTags);
+					if (sim.TagChanged) {
+						string datadir = Path.Combine(tmpdir, "data");
+						string bkpfile = recfile + ".bk";
 
-					Console.WriteLine("  -- old file was moved to " + bkpfile);
+						Console.WriteLine("  -- tags were modified, rewriting");
+						File.WriteAllText(Path.Combine(datadir, "tags.out"), sim.SerializedTags);
 
-					if (File.Exists(bkpfile)) {
-						File.Delete(bkpfile);
+						Console.WriteLine("  -- old file was moved to " + bkpfile);
+
+						if (File.Exists(bkpfile)) {
+							File.Delete(bkpfile);
+						}
+
+						File.Move(recfile, bkpfile);
+						
+						Console.WriteLine("  -- compressing");
+						ZipFile.CreateFromDirectory(datadir, recfile);
 					}
 
-					File.Move(recfile, bkpfile);
-					
-					Console.WriteLine("  -- compressing");
-					ZipFile.CreateFromDirectory(datadir, recfile);
+					Directory.Delete(tmpdir, true);
 				}
-
-				Directory.Delete(tmpdir, true);
+				else {
+					Console.WriteLine("Nothing to show! Make sure there are appropriate frames " +
+					                  "in the specified file. This can happen if screenshot mode is " +
+					                  "used and there are no screenshot tags.");
+				}
 			}
 		}
 		else {
 			using (var sim = Simulation<MeasurerT, PoseT, MeasurementT>.
 			                     FromFiles(scenefile, commandfile, particlecount, input,
 			                               algorithm, onlymapping, realtime, !headless, noterminate)) {
-				sim.CheckpointFile = recfile;
+				sim.ScreenshotPrefix = recfile + "-screenshot-";
+				sim.CheckpointFile   = recfile;
 				setSignalHandler(sim);
 
 				if (headless) {
